+++ src/inspect_ai/agent/_bridge/anthropic_api_impl.py
@@ -155,7 +155,13 @@ async def inspect_anthropic_api_request_impl(
 
     # if there is a bridge filter give it a shot first
     output, c_message = await bridge_generate(
-        bridge, model, messages, tools, tool_choice, config
+        bridge,
+        model,
+        messages,
+        tools,
+        tool_choice,
+        config,
+        requested_model=bridge_model_name,
     )
     if c_message is not None:
         messages.append(c_message)